correction

+++ binx/setup/setup.sh
@@ -1,27 +1,48 @@
 #!/bin/bash
 
 SSH_KEY_DUMP=$1
+TMPDIR=~/setuptmp
 
 function main()
 {
+    mkdir $TMPDIR
     cd ~/
     setup_apt_packages
     setup_ecrypt_private
-    setup_ssh_keys
+    if ! ssh-add -l
+    then
+	if [ "x$SSH_KEY_DUMP" = "x" ]
+	then
+	    echo ssh key encrypted dump no provided >&2
+	    exit -1
+	fi
+	setup_ssh_keys "$SSH_KEY_DUMP"
+    fi
+
+    if ! ssh-add -l
+    then
+	echo ssh key no available >&2
+	exit -1
+    fi
     setup_git_repos
 
     setup_user_config_setup
+
+    setup_download_misc
+
+    rm -rf $TMPDIR
 }
 
 function set_keyboard()
 {
-    wget 'https://raw.githubusercontent.com/sharad/rc/master/keymaps/Xmodmaps/xmodmaprc-swap-alt-ctrl-caps=alt' -O /tmp/keymap
-    xmodmap /tmp/keymap
+    mkdir $TMPDIR
+    wget 'https://raw.githubusercontent.com/sharad/rc/master/keymaps/Xmodmaps/xmodmaprc-swap-alt-ctrl-caps=alt' -O $TMPDIR/keymap
+    xmodmap $TMPDIR/keymap
 }
 
 function setup_apt_packages()
 {
-    : sudo apt install git-all ecryptfs-utils openssl stow sbcl cl-clx-sbcl
+    sudo apt install git-all ecryptfs-utils openssl stow sbcl cl-clx-sbcl at
 }
 
 function setup_ecrypt_private()
@@ -38,16 +59,17 @@ function setup_ecrypt_private()
 
 function setup_ssh_keys()
 {
+    SSH_KEY_ENC_DUMP=$1
     ## bring the ssh keys
-    if [ ! -r ~/tmp/ssh/nosecure.d/ssh/keys.d/github ]
+    if [ ! -r $TMPDIR/ssh/nosecure.d/ssh/keys.d/github ]
     then
-	mkdir -p ~/tmp/ssh
-	openssl enc -in "$SSH_KEY_DUMP" -aes-256-cbc -d | tar -zxvf - -C ~/tmp/ssh
+	mkdir -p $TMPDIR/ssh
+	openssl enc -in "$SSH_KEY_ENC_DUMP" -aes-256-cbc -d | tar -zxvf - -C $TMPDIR/ssh
     fi
 
     if ! ssh-add -l
     then
-	ssh-add ~/tmp/ssh/nosecure.d/ssh/keys.d/github
+	ssh-add $TMPDIR/ssh/nosecure.d/ssh/keys.d/github
     fi
 }
 
@@ -138,6 +160,17 @@ function setup_git_repos()
 	    ln -s ../.repos/git/user/orgp ~/.pi/org
 	fi
     fi
+
+    if [ ! -d ~/.repos/git/user/opt ]
+    then
+	git clone git@bitbucket.org:sh4r4d/opt.git ~/.repos/git/user/opt
+    fi
+
+    if [ ! -L ~/.opt ]
+    then
+	rm -rf ~/.opt
+	ln -sf ~/.repos/git/user/opt ~/.opt
+    fi
 }
 
 function setup_user_config_setup()
@@ -145,15 +178,44 @@ function setup_user_config_setup()
     if [ -d ~/.setup ]
     then
 	mkdir -p ~/.old_dot_filedirs
-	mv ~/.[a-zA-Z]* ~/.old_dot_filedirs/
-	for d in .repos .osetup .setup .setup-trunk .sysinfo .system .pi
+
+	mv ~/.setup/_home/.setup $TMPDIR/Xsetup
+	cd ~/.setup/_home/
+	for c in .[a-zA-Z]* *
 	do
-	    echo mv ~/.old_dot_filedirs/$d ~/
-	    mv ~/.old_dot_filedirs/$d ~/
+	    if [ "$c" != ".setup" ]
+	    then
+		if [ -e ~/$c ]
+		then
+		    mv ~/$c ~/.old_dot_filedirs
+		    cp -a $c ~/$c
+		fi
+	    fi
 	done
-	mv ~/.setup/_home/.setup ~/.setup/_home/.setup_x
-	cp -af ~/.setup/_home/.[a-zA-Z]* ~/
-	mv ~/.setup/_home/.setup_x ~/.setup/_home/.setup 
+	mv $TMPDIR/Xsetup ~/.setup/_home/.setup
+	cd ~/
+
+	for d in .repos .osetup .setup .setup-trunk .sysinfo .system .pi .config .gconf .ecryptfs
+	do
+	    if [ -e ~/.old_dot_filedirs/$d ]
+	    then
+		if [ ! -e ~/$d ]
+		then
+		    echo mv ~/.old_dot_filedirs/$d ~/
+		    mv ~/.old_dot_filedirs/$d ~/
+		fi
+	    fi
+	done
+    fi
+}
+
+function setup_download_misc()
+{
+    if [ ! -f /usr/local/bin/p4 ]
+    then
+	wget 'https://www.perforce.com/downloads/free/p4' -O $TMPDIR/p4
+	sudo cp $TMPDIR/p4 /usr/local/bin/p4
+	sudo chmod +x /usr/local/bin/p4
     fi
 }
